Make Glsl immutable
Before, the Glsl transforms list was mutated by every chained transform
call. Now, it's immutable, and each chained call will return a new
instance. This is implemented with a simple and efficient ImmutableList.

This is nice to have so you can share and extend the same transforms
between multiple outputs! And it protects noob like me from weird state
bugs.